+++ src/main/java/ru/mail/polis/lsm/DAO.java
@@ -21,12 +21,6 @@ public interface DAO extends Closeable {
 
     void compact();
 
-    /**
-     * Appends {@code Byte.MIN_VALUE} to {@code buffer}.
-     *
-     * @param buffer original {@link ByteBuffer}
-     * @return copy of {@code buffer} with {@code Byte.MIN_VALUE} appended
-     */
     static ByteBuffer nextKey(ByteBuffer buffer) {
         ByteBuffer result = ByteBuffer.allocate(buffer.remaining() + 1);
 
@@ -40,100 +34,4 @@ public interface DAO extends Closeable {
 
         return result;
     }
-
-    static Iterator<Record> merge(List<Iterator<Record>> iterators) {
-        if (iterators.size() == 0) {
-            return Collections.emptyIterator();
-        }
-        if (iterators.size() == 1) {
-            return iterators.get(0);
-        }
-        if (iterators.size() == 2) {
-            return mergeTwo(new PeekingIterator(iterators.get(0)), new PeekingIterator(iterators.get(1)));
-        }
-        Iterator<Record> left = merge(iterators.subList(0, iterators.size() / 2));
-        Iterator<Record> right = merge(iterators.subList(iterators.size() / 2, iterators.size()));
-        return mergeTwo(new PeekingIterator(left), new PeekingIterator(right));
-    }
-
-    static Iterator<Record> mergeTwo(PeekingIterator left, PeekingIterator right) {
-        return new Iterator<>() {
-
-            @Override
-            public boolean hasNext() {
-                return left.hasNext() || right.hasNext();
-            }
-
-            @Override
-            public Record next() {
-                if (!hasNext()) {
-                    throw new NoSuchElementException("No elements");
-                }
-
-                if (!left.hasNext()) {
-                    return right.next();
-                }
-                if (!right.hasNext()) {
-                    return left.next();
-                }
-
-                ByteBuffer leftKey = left.peek().getKey();
-                ByteBuffer rightKey = right.peek().getKey();
-
-                int compareResult = leftKey.compareTo(rightKey);
-                if (compareResult == 0) {
-                    left.next();
-                    return right.next();
-                }
-
-                if (compareResult < 0) {
-                    return left.next();
-                } else {
-                    return right.next();
-                }
-            }
-
-        };
-    }
-
-    class PeekingIterator implements Iterator<Record> {
-
-        private Record current;
-
-        private final Iterator<Record> delegate;
-
-        public PeekingIterator(Iterator<Record> delegate) {
-            this.delegate = delegate;
-        }
-
-        @Override
-        public boolean hasNext() {
-            return current != null || delegate.hasNext();
-        }
-
-        @Override
-        public Record next() {
-            if (!hasNext()) {
-                throw new NoSuchElementException();
-            }
-            Record now = peek();
-            current = null;
-            return now;
-        }
-
-        public Record peek() {
-            if (current != null) {
-                return current;
-            }
-
-            if (!delegate.hasNext()) {
-                return null;
-            }
-
-            current = delegate.next();
-            return current;
-        }
-
-    }
-
 }
