+++ src/main/java/ru/mail/polis/lsm/saveliyschur/utils/UtilsIterator.java
@@ -1,0 +1,65 @@
+package ru.mail.polis.lsm.saveliyschur.utils;
+
+import ru.mail.polis.lsm.Record;
+
+import java.nio.ByteBuffer;
+import java.util.Collections;
+import java.util.Iterator;
+import java.util.List;
+import java.util.NoSuchElementException;
+
+public class UtilsIterator {
+    public static Iterator<Record> merge(List<Iterator<Record>> iterators) {
+        if (iterators.size() == 0) {
+            return Collections.emptyIterator();
+        }
+        if (iterators.size() == 1) {
+            return iterators.get(0);
+        }
+        if (iterators.size() == 2) {
+            return mergeTwo(new PeekingIterator(iterators.get(0)), new PeekingIterator(iterators.get(1)));
+        }
+        Iterator<Record> left = merge(iterators.subList(0, iterators.size() / 2));
+        Iterator<Record> right = merge(iterators.subList(iterators.size() / 2, iterators.size()));
+        return mergeTwo(new PeekingIterator(left), new PeekingIterator(right));
+    }
+
+    public static Iterator<Record> mergeTwo(PeekingIterator left, PeekingIterator right) {
+        return new Iterator<>() {
+
+            @Override
+            public boolean hasNext() {
+                return left.hasNext() || right.hasNext();
+            }
+
+            @Override
+            public Record next() {
+                if (!hasNext()) {
+                    throw new NoSuchElementException("No elements");
+                }
+
+                if (!left.hasNext()) {
+                    return right.next();
+                }
+                if (!right.hasNext()) {
+                    return left.next();
+                }
+
+                ByteBuffer leftKey = left.peek().getKey();
+                ByteBuffer rightKey = right.peek().getKey();
+
+                int compareResult = leftKey.compareTo(rightKey);
+                if (compareResult == 0) {
+                    left.next();
+                    return right.next();
+                }
+
+                if (compareResult < 0) {
+                    return left.next();
+                } else {
+                    return right.next();
+                }
+            }
+        };
+    }
+}
